refactor: App.vue
