+++ .env.example
@@ -1,2 +1,2 @@
-VITE_APP_API_URL=https://my-json-server.typicode.com/ecanquiz/vue-todo-pwa
-
+VITE_SUPABASE_URL="https://example.supabase.co"
+VITE_SUPERBASE_KEY="<your_key>"
