+++ src/config/supabase.ts
@@ -1,7 +1,4 @@
 const supabaseUrl = import.meta.env.VITE_SUPABASE_URL; 
 const supabaseKey = import.meta.env.VITE_SUPABASE_KEY;
 
-export {
-    supabaseUrl,
-    supabaseKey
-}
+export { supabaseUrl, supabaseKey }
